feat：集成mybatis plus商品分类接口基本完成

+++ src/main/java/com/shoppingonline/pojo/ProductType.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class ProductType {
-    private Integer id;//主键ID
+    private Long id;//主键ID
     private String name;//分类名称
     private String description;//分类描述
     private String iconName;//分类图标
